fix #11: stack iframe and transparent webgl
Iframe and the transparrent webgl window is now stacked on top of each
other. Done by modifying the .html document that the application resides
in. Currently all information is received from the "gameContainer" tag,
which ensures that the iframe window will be of correct size also when we
change the window resolution during startup.

Currently the webgl window is not fully transparrent, as we probably want
to see it when testing/developing.